+++ src/components/ProfileCard.vue
@@ -7,7 +7,7 @@
       <div class="avatar-row">
         <avatar :profile="profile"></avatar>
         <div class="name-group">
-          <div class="display-name">{{ profile.display_name || profile.username }}</div>
+          <div class="display-name">{{ profile.getDisplayName() }}</div>
           <div class="actor-address">@{{ getActorAddress(profile) }}</div>
         </div>
       </div>
@@ -23,17 +23,21 @@
 </template>
 
 <script setup lang="ts">
-import { Profile } from "@/api/users"
+import { $computed } from "vue/macros"
+
+import { Profile, ProfileWrapper } from "@/api/users"
 import Avatar from "@/components/Avatar.vue"
 import { useInstanceInfo } from "@/store/instance"
 
 /* eslint-disable-next-line no-undef */
-defineProps<{
+const props = defineProps<{
   profile: Profile,
   compact: boolean,
 }>()
 
 const { getActorAddress } = useInstanceInfo()
+
+const profile = $computed(() => new ProfileWrapper(props.profile))
 </script>
 
 <style scoped lang="scss">